Fix cmake warning regarding module name (#79)
cmake 3.19 prints a warning if variables set from a "find module" does not
follow the normal pattern, i.e. a find module named FindUSB1.cmake is assumed
to set variables that begin with USB1_*. This patch renames the module to be
consistent with the names of the variables set from inside it.